+++ packages/components/ContextualMenu/src/ContextualMenu.settings.ts
@@ -1,5 +1,6 @@
-import { contextualMenuName, ContextualMenuType } from './ContextualMenu.types';
-import { IComposeSettings } from '@uifabricshared/foundation-compose';
+import type { ContextualMenuType } from './ContextualMenu.types';
+import { contextualMenuName } from './ContextualMenu.types';
+import type { IComposeSettings } from '@uifabricshared/foundation-compose';
 
 export const settings: IComposeSettings<ContextualMenuType> = [
   {