+++ packages/experimental/ActivityIndicator/src/ActivityIndicator.styling.tsx
@@ -1,13 +1,15 @@
-import { UseStylingOptions, buildProps } from '@fluentui-react-native/framework';
-import { ActivityIndicatorProps as CoreActivityIndicatorProps, Appearance } from 'react-native';
-import {
-  activityIndicatorName,
+import type { UseStylingOptions } from '@fluentui-react-native/framework';
+import { buildProps } from '@fluentui-react-native/framework';
+import type { ActivityIndicatorProps as CoreActivityIndicatorProps } from 'react-native';
+import { Appearance } from 'react-native';
+import type {
   ActivityIndicatorProps,
   FluentActivityIndicatorSlotProps,
   ActivityIndicatorTokens,
   CoreActivityIndicatorSlotProps,
   ActivityIndicatorSize,
 } from './ActivityIndicator.types';
+import { activityIndicatorName } from './ActivityIndicator.types';
 import assertNever from 'assert-never';
 
 export const diameterSizeMap: { [key: string]: number } = {