+++ packages/experimental/AppearanceAdditions/src/getSizeClass.ios.ts
@@ -4,7 +4,7 @@ import { useSubscription } from 'use-subscription';
 import { appearanceAdditions } from './appearanceAdditions';
 
 import NativeAppearanceAdditions from './NativeAppearanceAdditions';
-import { SizeClass } from './NativeAppearanceAdditions.types';
+import type { SizeClass } from './NativeAppearanceAdditions.types';
 
 const eventEmitter = NativeAppearanceAdditions ? new NativeEventEmitter(NativeAppearanceAdditions as any) : undefined;
 
@@ -13,9 +13,11 @@ export function useHorizontalSizeClass(): SizeClass {
     return 'regular';
   }
 
+  // Early return on eventEmitter will either always or never return within a single instance
+  // eslint-disable-next-line react-hooks/rules-of-hooks
   const subscription = useMemo(
     () => ({
-      getCurrentValue: () => appearanceAdditions.horizontalSizeClass,
+      getCurrentValue: () => appearanceAdditions().horizontalSizeClass,
       subscribe: (callback) => {
         const appearanceSubscription = eventEmitter.addListener('appearanceChanged', callback);
         return () => {
@@ -26,5 +28,7 @@ export function useHorizontalSizeClass(): SizeClass {
     [],
   );
 
+  // Early return on eventEmitter will either always or never return within a single instance
+  // eslint-disable-next-line react-hooks/rules-of-hooks
   return useSubscription(subscription);
 }