+++ packages/experimental/Spinner/src/Spinner.tsx
@@ -2,8 +2,10 @@
 import { View } from 'react-native';
 import { TextV1 as Text } from '@fluentui-react-native/text';
 import { Svg } from 'react-native-svg';
-import { compose, mergeProps, withSlots, UseSlots } from '@fluentui-react-native/framework';
-import { spinnerName, SpinnerProps, SpinnerType } from './Spinner.types';
+import type { UseSlots } from '@fluentui-react-native/framework';
+import { compose, mergeProps, withSlots } from '@fluentui-react-native/framework';
+import type { SpinnerProps, SpinnerType } from './Spinner.types';
+import { spinnerName } from './Spinner.types';
 import { RCTNativeAnimatedSpinner } from './consts.win32';
 
 /* TODO: Implement Spinner with following slots */