+++ packages/framework/theme/src/mergeTheme.ts
@@ -1,5 +1,5 @@
 import { immutableMerge } from '@fluentui-react-native/immutable-merge';
-import { Theme, PartialTheme } from '@fluentui-react-native/theme-types';
+import type { Theme, PartialTheme } from '@fluentui-react-native/theme-types';
 
 export function mergeTheme<TTheme extends Theme, TPartial extends PartialTheme>(base: TTheme, partial: TPartial): TTheme {
   return immutableMerge(base, partial as unknown as TTheme);