+++ packages/theming/win32-theme/src/getThemeTypography.ts
@@ -1,4 +1,4 @@
-import { Typography, Variants } from '@fluentui-react-native/theme-types';
+import type { Typography, Variants } from '@fluentui-react-native/theme-types';
 import { defaultFluentTheme } from '@fluentui-react-native/default-theme';
 import { createFontAliasTokens } from './createFontAliasTokens';