+++ packages/experimental/Spinner/CHANGELOG.md
@@ -1,8 +1,37 @@
 # Change Log - @fluentui-react-native/spinner
 
-This log was last generated on Wed, 25 Jan 2023 21:31:18 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Feb 2023 01:29:38 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.5
+
+Thu, 02 Feb 2023 01:29:38 GMT
+
+### Patches
+
+- Add eslint to packages missing a config (30809111+acoates-ms@users.noreply.github.com)
+- Bump @fluentui-react-native/framework to v0.8.39
+- Bump @fluentui-react-native/text to v0.19.18
+
+## 0.1.4
+
+Wed, 01 Feb 2023 22:33:17 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/component-cache to v1.4.3
+- Bump @fluentui-react-native/framework to v0.8.38
+- Bump @fluentui-react-native/text to v0.19.17
+
+## 0.1.3
+
+Mon, 30 Jan 2023 07:12:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.37
+- Bump @fluentui-react-native/text to v0.19.16
 
 ## 0.1.2