+++ packages/framework/framework/CHANGELOG.md
@@ -1,8 +1,50 @@
 # Change Log - @fluentui-react-native/framework
 
-This log was last generated on Wed, 25 Jan 2023 21:31:18 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Feb 2023 01:29:39 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.39
+
+Thu, 02 Feb 2023 01:29:39 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/composition to v0.8.3
+- Bump @fluentui-react-native/default-theme to v0.16.29
+- Bump @fluentui-react-native/use-styling to v0.9.3
+- Bump @fluentui-react-native/use-slot to v0.3.3
+- Bump @fluentui-react-native/use-slots to v0.7.3
+- Bump @fluentui-react-native/use-tokens to v0.3.3
+
+## 0.8.38
+
+Wed, 01 Feb 2023 22:33:15 GMT
+
+### Patches
+
+- Enable @typescript-eslint/consistent-type-imports (30809111+acoates-ms@users.noreply.github.com)
+- Bump @fluentui-react-native/composition to v0.8.2
+- Bump @fluentui-react-native/default-theme to v0.16.28
+- Bump @fluentui-react-native/immutable-merge to v1.1.8
+- Bump @fluentui-react-native/merge-props to v0.5.2
+- Bump @fluentui-react-native/theme-types to v0.30.2
+- Bump @fluentui-react-native/use-styling to v0.9.2
+- Bump @fluentui-react-native/use-slot to v0.3.2
+- Bump @fluentui-react-native/use-slots to v0.7.2
+- Bump @fluentui-react-native/use-tokens to v0.3.2
+- Bump @fluentui-react-native/memo-cache to v1.1.8
+- Bump @fluentui-react-native/tokens to v0.20.7
+
+## 0.8.37
+
+Mon, 30 Jan 2023 07:12:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.16.27
+- Bump @fluentui-react-native/theme-types to v0.30.1
+- Bump @fluentui-react-native/tokens to v0.20.6
 
 ## 0.8.36