+++ packages/theming/android-theme/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @fluentui-react-native/android-theme
 
-This log was last generated on Thu, 26 Jan 2023 18:46:08 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Feb 2023 01:29:39 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.16.4
+
+Thu, 02 Feb 2023 01:29:39 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme to v0.7.21
+
+## 0.16.3
+
+Wed, 01 Feb 2023 22:33:14 GMT
+
+### Patches
+
+- Enable @typescript-eslint/consistent-type-imports (30809111+acoates-ms@users.noreply.github.com)
+- Bump @fluentui-react-native/theme-tokens to v0.24.2
+- Bump @fluentui-react-native/theme-types to v0.30.2
+- Bump @fluentui-react-native/theme to v0.7.20
+- Bump @fluentui-react-native/theming-utils to v0.23.2
+- Bump @fluentui-react-native/memo-cache to v1.1.8
+
+## 0.16.2
+
+Mon, 30 Jan 2023 07:12:38 GMT
+
+### Patches
+
+- Add `darkElevated` theme (mischreiber@microsoft.com)
+- Bump @fluentui-react-native/theme-tokens to v0.24.1
+- Bump @fluentui-react-native/theme-types to v0.30.1
+- Bump @fluentui-react-native/theme to v0.7.19
+- Bump @fluentui-react-native/theming-utils to v0.23.1
 
 ## 0.16.1